refactor: modularize pool and worker ORM
